finish basic Django Project

+++ BootCampTask/Django/my_first_pjt/articles/views.py
@@ -5,6 +5,7 @@ https://docs.djangoproject.com/en/4.2/
 from django.shortcuts import get_object_or_404, redirect, render
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_POST, require_http_methods
+from django.contrib.auth import get_user_model
 
 from .forms import ArticleForm, CommentForm
 from .models import Article, Comment
@@ -20,7 +21,9 @@ def create(request):
     form = ArticleForm(request.POST, request.FILES) if request.method == "POST" else ArticleForm()
     
     if request.method == "POST" and form.is_valid():
-        article = form.save()
+        article = form.save(commit=False)
+        article.author = request.user
+        article.save()
         return redirect("articles:article_detail", article.id)
 
     context = {"form": form}
@@ -40,37 +43,42 @@ def article_detail(request, pk):
         "article" : article, 
         "comment_form" : comment_form,
         "comments" : comments,
-        "total_comments" : total_comments
+        "total_comments" : total_comments,
         }
     return render(request, "articles/article_detail.html", context)
 
+@login_required
 @require_POST
 def delete(request, pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=pk)
-        article.delete()
+        if request.user == article.author:
+            article.delete()
     return redirect("articles:articles")
 
 @login_required
 @require_http_methods(["GET", "POST"])
 def update(request, pk):
     article = get_object_or_404(Article, pk=pk)
-    
-    if request.method == "POST":
-        # instance를 선언하면, 새로 만드는게 아닌 기존것을 수정한다다
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            article = form.save()
-            return redirect("articles:article_detail", article.pk)
+    if request.user == article.author:
+        if request.method == "POST":
+            # instance를 선언하면, 새로 만드는게 아닌 기존것을 수정한다
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                article = form.save()
+                return redirect("articles:article_detail", article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
-
+        return redirect("articles:articles")
+    
     context = {
         "form": form,
         "article": article,
     }
     return render(request, "articles/update.html", context)
 
+@login_required
 @require_POST
 def comment_create(request, pk):
     article = get_object_or_404(Article, pk=pk)
@@ -78,16 +86,29 @@ def comment_create(request, pk):
     if form.is_valid():
         comment = form.save(commit=False)
         comment.article = article
+        comment.author = request.user
         comment.save()
     return redirect("articles:article_detail", article.pk)
 
+@login_required
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    if request.user == comment.author:
+        comment.delete()
     return redirect("articles:article_detail", article_pk)
 
-
+@require_POST
+def like(request, article_pk):
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        user = request.user
+        if article.like_users.filter(pk=user.pk).exists():
+            article.like_users.remove(user)
+        else:
+            article.like_users.add(user)
+        return redirect("articles:articles")
+    return redirect("accounts:login")
